+++ test/bigtest/network/factories/campus.js
@@ -1,0 +1,11 @@
+
+import {
+  Factory,
+  faker,
+} from '@bigtest/mirage';
+
+export default Factory.extend({
+  id: faker.random.uuid,
+  name: faker.hacker.noun(),
+  code: faker.hacker.abbreviation(),
+});